Produce valid Haml for poorly indented filters
When a css or javascript filter is being generated, if the first line of
the code in the source is indented more than any other line then invalid
Haml is produced.

Fix this by simply stripping the leading whitespace away, and indenting
all lines by the current indentation amount. This may result in odd
looking output, but will at least be valid Haml.